Move RemoteStorageConfig, RemoteStorageKind, AzureConfig and S3Config to own module
